effec ted changes from codereview
