+++ enumerable_methods1.rb
@@ -35,7 +35,7 @@ module Enumerable
     if !pattern.nil?
       obj.my_each { |i| all = false unless pattern === i }
     elsif !block_given? && pattern.nil?
-      obj.my_each { |i| all = false if i == (false || nil) }
+      obj.my_each { |i| all = false if i == false || i == nil }
     else
       obj.my_each { |i| all = false unless yield i }
     end
@@ -89,6 +89,54 @@ module Enumerable
     obj.my_each do |i|
       item = yield(i)
       my_map_array << item
+    end
+    my_map_array
+  end
+
+  def my_inject(initial = 0, opr = nil)
+    my_array = []
+    if self.class == Range
+      for i in self
+        my_array << i
+      end
+      obj = my_array
+    else
+        obj = self
+    end
+    if initial.class == (Symbol or String)
+      opr = initial.to_s
+      (obj.size - 1).times do |i|
+        obj[i.next] = obj[i].send(opr, obj[i.next])
+        initial = obj[i.next]
+      end
+    elsif opr.class == (Symbol or String)
+      obj.size.times { |i| initial = initial.send(opr.to_s, obj[i]) }
+    elsif initial != 0
+      obj.size.times { |i| initial = yield(initial, obj[i]) }
+    else
+      (obj.size - 1).times do |i|
+        obj[i.next] = yield(obj[i], obj[i.next])
+        initial = obj[i.next]
+      end
+    end
+    initial
+  end
+
+  def my_map(&proc)
+    return to_enum unless block_given?
+
+    my_map_array = []
+    obj = self
+    if block_given?
+      obj.my_each do |i|
+        item = yield(i)
+        my_map_array << item
+      end
+    else
+      obj.my_each do |i|
+        item = proc.call(i)
+        my_map_array << item
+      end
     end
     my_map_array
   end
